Use a real response from the facetec server at 3D DB enroll
For some reason, our FaceTec server doesn't provide
the externalDatabaseRefID field in the response.
The test server and postman test suite both do provide it.
We don't really need it, so it's simply removed from the response,
but it looks like we can't trust any of the FaceTec docs / test
servers.