Work around validator key availability problem
The validator public key is not available until after the rpc handlers
are initialized. This means we can't pass it to bioauth before then. These
changes switch back to the validator key extractor, which gets the
key after its initialized. A validator signer factory is also necessary,
as a signer can only be created once we know the key.